Move baseline utils from common to a dedicated utils folder (#482)
This is to move habitat_baselines utils from habitat_baselines/common to a dedicated habitat_baselines/util
