+++ habitat_baselines/utils/__init__.py
@@ -4,4 +4,4 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-__all__ = ["visualizations"]
+__all__ = ["visualizations", "env_utils", "common"]
